Moves FormFieldToggleSwitch
